refactor: reduce maximum connections each pod can service (#75)
With 10,000 connections and a block batch limit of 512 we _could_, in the worst case scenario be asked for `10,000 * 512 * 1MB` of blocks concurrently which would require ~`5TB` of memory! More normally (perhaps - guestimated), each peer might have 3MB of data outstanding, which would still require `30GB` of memory.

This PR reduces the maximum number of connections to 1,000 so that in the above normal case we use `3GB` of the `6GB` of available memory, with room for bursts.