chore: ensure we have at least 50 pods running (#97)
our autoscaling rules are not yet calibrated. we need to explore scaling based on node event loop saturation as cpu usage is pretty steady regardless while mem usage is too spikey to be a useful indicator. the event loop utilization hitting 100% seems to corrolate more closely with when pods start struggling.

in the meantime lets ensure we have enough pods around and make full use of the instances we have

License: MIT

Signed-off-by: Oli Evans <oli@protocol.ai>